api/requirements.txt
Flask
flask-cors
flask-mysql
mysql-connector-python
simplejson
